Split part of UI to TweetView.qml
Add unit tests to it. Not very thorough test, though. Having problems
with the visible-property in QmlUnit tests, it seems that it's always
false even though it clearly works when really rendered on the UI.
Either there's some funny little thing I'm just missing, or there's some
kind of a bug/missing feature in Qml or QmlUnit that's preventing this
from working.

Signed-off-by: Aki Saarinen <aki@akisaarinen.fi>